Configurable thingies for testing, separate files a bit.
